requirements.txt
requests
pynput
Pillow
python-dotenv
